% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_silhouette.R
\name{fast.sil}
\alias{fast.sil}
\title{Calculating Silhouette on Linear Data Clusters}
\usage{
fast.sil(x, cluster)
}
\arguments{
\item{x}{a numeric vector of one-dimensional points}

\item{cluster}{an integer vector of cluster labels for each point}
}
\value{
The function returns a numeric value of the average
 silhouette information calculated on the input data clusters.
}
\description{
A fast linear-time algorithm to calculate silhouette information on
one-dimensional data with cluster labels.
}
\details{
The silhouette information on one-dimensional data is
calculated in linear time here, instead of quadratic time by
definition. There is an overhead of sorting \eqn{O(n \log n)}{O(n log n)} if the
input data are not sorted.
}
\examples{
x <- c(-1.2, -2, -3, -2.5, 1, 0.8, 1.5, 1.2)
cluster <- c(1, 1, 1, 1, 2, 2, 2, 2)
fast.sil(x, cluster)


}
